Organize colors styling
