+++ frontend/app/index.tsx
@@ -1,7 +1,24 @@
-import { View } from "react-native";
+import { StyleSheet, View } from "react-native";
 import MapView from "react-native-maps";
-import styles from "./styles";
 import Search from "./components/search/Search";
+
+const styles = StyleSheet.create({
+  container: {
+      ...StyleSheet.absoluteFillObject,
+      justifyContent: 'center',
+      alignItems: 'center',
+  },
+  map: {
+      ...StyleSheet.absoluteFillObject,
+  },
+  search: {
+      ...StyleSheet.absoluteFillObject,
+      position: 'absolute',
+      top: 54,
+      left: 16.5,
+      right: 16.5,
+  }
+});
 
 export default function Index() {
   return (
