+++ frontend/app/styles/colors.tsx
@@ -1,0 +1,12 @@
+export const background = {
+    primary: "white",
+    secondary: "lightgray",
+};
+
+export const text = {
+    primary: "black",
+};
+
+export const underlay = {
+    primary: "#eeeeee",
+};
