+++ frontend/app/styles/index.tsx
@@ -1,0 +1,1 @@
+export * as Colors from './colors';
